updated proposal details page and added some more tests
